First rudimentary implementation
